+++ src/Symfony/Component/Mime/Test/Constraint/EmailHtmlBodyContains.php
@@ -1,0 +1,56 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Mime\Test\Constraint;
+
+use PHPUnit\Framework\Constraint\Constraint;
+
+final class EmailHtmlBodyContains extends Constraint
+{
+    private $expectedText;
+
+    public function __construct(string $expectedText)
+    {
+        $this->expectedText = $expectedText;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function toString(): string
+    {
+        return sprintf('contains "%s"', $this->expectedText);
+    }
+
+    /**
+     * @param RawMessage $message
+     *
+     * {@inheritdoc}
+     */
+    protected function matches($message): bool
+    {
+        if (RawMessage::class === \get_class($message) || Message::class === \get_class($message)) {
+            throw new \LogicException('Unable to test a message HTML body on a RawMessage or Message instance.');
+        }
+
+        return false !== mb_strpos($message->getHtmlBody(), $this->expectedText);
+    }
+
+    /**
+     * @param RawMessage $message
+     *
+     * {@inheritdoc}
+     */
+    protected function failureDescription($email): string
+    {
+        return 'the Email HTML body '.$this->toString();
+    }
+}