+++ src/Symfony/Component/Mime/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 4.4.0
 -----
 
+ * Added PHPUnit constraints
  * Added `AbstractPart::asDebugString()`
 
 4.3.3